requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
setuptools==73.0.1
sqlparse==0.5.1
uritemplate==4.1.1
